Island: Move AuthenticationService to authentication_service dir
