+++ monkey/monkey_island/cc/resources/auth/registration_status.py
@@ -1,9 +1,8 @@
 from monkey_island.cc.flask_utils import AbstractResource
-from monkey_island.cc.services import AuthenticationService
+from monkey_island.cc.services.authentication_service import AuthenticationService
 
 
 class RegistrationStatus(AbstractResource):
-
     urls = ["/api/registration-status"]
 
     def __init__(self, authentication_service: AuthenticationService):
