+++ monkey/monkey_island/cc/services/__init__.py
@@ -1,5 +1,4 @@
 from .agent_signals_service import AgentSignalsService
-from .authentication_service import AuthenticationService
 
 from .aws import AWSService
