+++ monkey/monkey_island/cc/services/authentication_service/__init__.py
@@ -1,0 +1,1 @@
+from .authentication_service import AuthenticationService
