+++ monkey/monkey_island/cc/services/authentication_service/authentication_service.py
