+++ monkey/tests/unit_tests/monkey_island/cc/resources/auth/test_login.py
@@ -5,7 +5,7 @@ import pytest
 from flask import Response
 
 from monkey_island.cc.resources.auth import Login
-from monkey_island.cc.services import AuthenticationService
+from monkey_island.cc.services.authentication_service import AuthenticationService
 
 USERNAME = "test_user"
 PASSWORD = "test_password"
