+++ monkey/tests/unit_tests/monkey_island/cc/services/authentication_service/test_authentication_service.py
@@ -6,7 +6,7 @@ import pytest
 from monkey_island.cc.event_queue import IIslandEventQueue, IslandEventTopic
 from monkey_island.cc.models import IslandMode, User
 from monkey_island.cc.server_utils.encryption import ILockableEncryptor
-from monkey_island.cc.services import AuthenticationService
+from monkey_island.cc.services.authentication_service import AuthenticationService
 
 USERNAME = "user1"
 PASSWORD = "test"
